requirements.txt
torch
diffusers
transformers
safetensors
omegaconf
trimesh
Pillow
numpy
einops
realesrgan
basicsr
